add recipe module (remove Recipe from __init__.py)
