+++ src/plone/recipe/zeoserver/ctl.py
@@ -19,9 +19,9 @@ import sys
 
 
 if sys.platform[:3].lower() == "win":
-    print 'For win32 platforms, runzeo.bat or zeoservice.exe should be used'
-    print '%s is based on zdaemon, which is Linux specific' % sys.argv[0]
-    print 'Aborting...'
+    print('For win32 platforms, runzeo.bat or zeoservice.exe should be used')
+    print('%s is based on zdaemon, which is Linux specific' % sys.argv[0])
+    print('Aborting...')
     sys.exit(0)
